Address, name, logo: Store extraction date and cache timestamp
For address, logo, and name, we want to keep track of both the 
extraction date (last time the pipeline has run) and the cache date 
(when the API query was last run). This allows us to correctly identify 
when to re-run the pipeline after a new scrape (extraction date < 
scraping date) and after a certain staleness interval, when API queries 
need to be refreshed (extraction date + 30 days < now).