Pipeline web: Scrape up to 100 pages per domain
Set a per-domain limit of 100 pages. A couple domains of several
hundreds or even thousands of pages. We don't need to collect all of
them at this point where we're still prototyping our approach. Later,
when we have better computing power and can run several scrapers in
parallel, we can scrape the entire website.