Google Maps API: Query all combinations of the URL & domain
Queries to find a place on Google Maps are free as long as we only 
request the Places ID in response. So we can actually run various 
permutations of the url and domain (with quotation marks and without, 
with https:// and without, etc...) at no extra cost. This is necessary, 
because some places on Google Maps can only be found with one of these 
special configurations.