Pipeline logo: Upload to Cloudinary
Upload extracted logos to Cloudinary and store the public logo URL in 
our database. This way, the frontend and admin dashboard have access to 
the logos directly via URL and we don't have to manually transfer logo 
files from this repo to the other ones.

To avoid excessively uploading logos to Cloudinary, we hash the logo and 
store a logo_hash. When we next extract an image for the organization, 
we check if the hash is the same. If so, we skip the Cloudinary upload 
because the logo has not changed.