Webpage: Generate random file name when setting (new) content
Instead of relying on webpage ID for the file path, we generate a random
UUID whenever new content is set. This has two advantages: It is not
dependent on the record existing prior to setting content and it does
not require special logic for when a transaction is aborted or rolled
back as the old content remains available under the previous file_name.