Webpage: Store HTML content in database rather than externally
Initially, the Webpage model was designed to store its HTML content in
an external file with GZIP compression. This was based on the GGD v1
directory where we were using an SQLite database. Then, using external
files with GZIP compression brought massive disk space improvements
(reductions upwards of 50%). However, it does have some downsides, most
notably that a lot of logic is required to make external files
compatible with database transactions and to properly clean up external
files when a row is deleted.

PostgreSQL provides built-in compression for any text-based content over
2KB in size. This compression is not as effective as GZIP compression
(albeit faster), but should be good enough for now. And it allows us to
overcome the above-mentioned downsides.

For more info, see https://stackoverflow.com/questions/3801416/text-compression-in-postgresql