ExtractPipeline: Add URL blocklist and ignore matching pages
For some pages, we do not want to extract and analyze the content. For
example, for privacy policy pages or for news articles. To ignore these
pages, the ExtractPipeline has a URL_BLOCKLIST that is used to skip a
URL if any of the URL's path segments are contained in the blocklist.