TextBlock: Make website-scoped by adding website_id column
Instead of having globally unique text blocks, text blocks are now
scoped to websites and unique within each website. Each website can only
have one unique text block for the same piece of content. But for
several websites with the same piece of content, each website will have
its own 'copy' of the text block. This allows us to clear/delete all
text blocks for a website when we trigger the ExtractPipeline without
affecting any of the other websites and causing possible inconsistencies
(for example, between the keywords found and the SDG mapping).