Extract: Improve performance by caching block hashes and IDs
Speed up the extraction pipeline by caching text block hashes and the
associated ID of the text block record in a dict/hash map. When the
pipeline discovers a duplicate piece of content (with the same hash),
the associated text block ID can be found in the dict/hash map instead
of needing to query the database.