Socials: Identify main handles + email for each organization
After extracting socials (handles and email addresses), identify the one
handle (per social network) and the one email address that is most
likely the organization's main one. This identification is based on the
frequency with which each handle/email appears as well as how similar
the handle/email address is to the organization's domain name.