KeywordsPipeline: Add new "generic" keyword type
To better map organizations to the SDGs, we include some very broad
keywords in our keyword list (such as food for SDG 2). These
terms by themselves are too broad to clearly map to the relevant SDG
because they might often also be used in another context (as in the case
of food, which does not always refer to the topic of diets and
nutrition). If only generic keywords are detected, the organization is
not mapped as working on that SDG. But if at least one specific /
non-generic keyword is found, the generic keywords contribute to the
calculation of the keyword density.